requirements.txt
anthropic>=0.40.0
yfinance>=0.2.50
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
tabulate>=0.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
